fix: legacy setup code from Angular 15 to 17 migration
This commit attempts to fix legacy setup code from the Angular 15 to 17 migration. Since Angular 17 uses standalone mode instead of modules as default and the Angular documentation for v17 presumes standalone mode, it is difficult to percieve what the proper Angular 17 setup should be for modules based apps with both server side rendering and i18n enabled.

This commit:
- Changes `HttpClientModule` to `provideHttpClient(withFetch())` (recommended for SSR apps: https://angular.io/api/common/http/provideHttpClient).
- Changes the polyfills config in `angular.json` and deletes the `src/polyfills.ts` file.
- Renames the app component from `DigitalEditionApp` to the default `AppComponent`.
- Removes the check for `document.readyState` and the `DOMContentLoaded` event listener before bootstrapping in `main.ts`.

Most of the changes are based on what the source code of a new modules based Angular 17 app with SSR and i18n enabled looks like (generated with `ng new <app-name> --routing --ssr --no-standalone`).