feature: define intrinsic size of home page banner image
Added config option to define the intrinsic dimensions of the banner image on the home page: `config.page.home.bannerImage.intrinsicSize`. The option defaults to `null`, meaning that the `height` and `width` attributes are not set on the image. The option takes an object with `height` and `width` keys and their respective values as numbers (the implicit unit is pixels). Setting the intrinsic dimensions of the image is recommended since the browser can then calculate the aspect ratio of the image, which improves initial page rendering.