fix: correctly set document title on revisiting collection pages in nav stack
Prior to this fix the document title was not set correctly when going back to a collection page that had already been visited. This was because in Ionic, the components are not destroyed when leaving views, but saved in a nav stack. The fix uses the Ionic life cycle hooks ionViewWillEnter and ionViewWillLeave on the collection pages to keep track of if the component is active or not, and this information is passed to the text-changer component which sets the document title.

The text-changer component has been refactored to get the information about the current page from the router instead of from input variables.

The document head service has been optimized to avoid unnecessary DOM updates by setting the document title only if it changes.