feat(elasticsearch-page): support PDF type search results
Elasticsearch page: support PDF type search results for PDFs ingested in Elasticsearch as one page in plain text format per document. For links from the search results to correctly map to PDFs on the ebook pages, you need to enter the PDFs as collections in the database, and add a `collectionId` property to each ebook object in the ebooks array in `config.ts`.